+++ src/index.ts
@@ -35,6 +35,7 @@ export type { TypedContractMethod } from './token-registry-v5/typedContractMetho
 export type { CancelTransactionSigner } from './transaction';
 export * from './token-registry-functions';
 export * from './obligation-registry-functions';
+export * from './eip7702-functions';
 export * from './core';
 export * from './open-attestation';
 export * from './verify';